+++ lib/widgets/add_limit_button.dart
@@ -2,11 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:myapp/config/constants/colors.dart';
 
-
-// || HOW TO USE,
-// INPUTS: an onclick function
-// OUTPUTS: on button click outputs the onclick function
-
 class AddLimitButton extends StatelessWidget {
   final VoidCallback onPressed; // Function to handle button press.
