+++ lib/widgets/custom_button.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class CustomButton extends StatelessWidget {
-  final String text; // Button text.
-  final VoidCallback onPressed; // Function to handle button press.
+  final String text;
+  final VoidCallback onPressed;
 
   const CustomButton({super.key, required this.text, required this.onPressed});
